Add missing simulator architecture
