--- ios/PolyPodApp/Dependencies/PolyPodCore.xcframework/ios-arm64-simulator/Headers/module.modulemap
@@ -1,4 +1,0 @@
-module PolyPodCore {
-    header "polypodcore.h"
-    export *
-}
