+++ ios/PolyPodApp/Dependencies/PolyPodCore.xcframework/ios-arm64_x86_64-simulator/Headers/module.modulemap
@@ -1,0 +1,4 @@
+module PolyPodCore {
+    header "polypodcore.h"
+    export *
+}
